+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/cache_app_dsn.php
@@ -1,0 +1,8 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'cache' => [
+        'app' => 'redis://example.com:6380',
+        'system' => 'memcached://example.com:11211',
+    ],
+]);